requirements.txt
discord.py
Pillow
natsort
urllib3